requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
dnspython==2.6.1
future==1.0.0
idna==3.6
python-whois==0.8.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
